+++ test/abi_test.exs
@@ -1,4 +1,5 @@
 defmodule ABITest do
-  use ExUnit.Case
+  use ExUnit.Case, async: true
+  use ABI.Hex
   doctest ABI
 end